# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for SEIIKI electrical service operations
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryRecentActivityItem } from './dashboardSummaryRecentActivityItem';

export interface DashboardSummary {
  totalRequests: number;
  pendingAssignment: number;
  onSite: number;
  completed: number;
  visitRevenue: number;
  repairRevenue: number;
  recentActivity: DashboardSummaryRecentActivityItem[];
}