# lib/api-zod/src/generated/types/fieldReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for SEIIKI electrical service operations
 * OpenAPI spec version: 0.1.0
 */

export interface FieldReport {
  id: number;
  requestId: number;
  notes: string;
  media: string[];
  createdAt: Date;
}