# lib/api-zod/src/generated/types/serviceRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for SEIIKI electrical service operations
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceRequestInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 8 */
  whatsapp: string;
  /** @minLength 4 */
  address: string;
  latitude: number;
  longitude: number;
  serviceType: string;
  notes?: string;
}