# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for SEIIKI electrical service operations
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';
import type { UserUpdateStatus } from './userUpdateStatus';

export interface UserUpdate {
  name?: string;
  phone?: string;
  role?: UserUpdateRole;
  status?: UserUpdateStatus;
}